fix : NavbarModal
